+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Group invitation email template fix
 - Response loader for requests in Modals
 - Group Roles Filter in Group Members Tab
+- Updated Validation for Membership Expiration in Enrollment configuration to match Backend validation
+- Changed enrollment to enrolment in messages
+- Fixed Typos
 
 ## [1.6.1] - 2025-04-03